+++ Content/Items/Materials/BrokenSwordShards.cs
@@ -16,7 +16,7 @@ namespace SupernovaMod.Content.Items.Materials
 		{
 			Item.width = 16;
 			Item.height = 16;
-			Item.maxStack = 999;
+			Item.maxStack = 9999;
 			Item.rare = ItemRarityID.LightRed;
 			Item.value = Item.buyPrice(1);
 		}