+++ Content/Items/Materials/EldritchEssence.cs
@@ -21,6 +21,7 @@ namespace SupernovaMod.Content.Items.Materials
 			Item.height = 16;
 			Item.maxStack = 9999;
 			Item.rare = ItemRarityID.Pink;
+			Item.value = Item.sellPrice(0, 0, 80);
 		}
 	}
 }