+++ Content/Items/Materials/HelixStone.cs
@@ -16,9 +16,9 @@ namespace SupernovaMod.Content.Items.Materials
         {
             Item.width = 16;
 			Item.height = 16;
-			Item.maxStack = 999;
+			Item.maxStack = 9999;
 			Item.rare = ItemRarityID.LightRed;
-			Item.value = Item.buyPrice(0, 1, 0, 0);
+			Item.value = Item.buyPrice(0, 1, 20, 0);
         }
     }
 }