+++ Content/Items/Materials/TitaniumRingMold.cs
@@ -17,7 +17,7 @@ namespace SupernovaMod.Content.Items.Materials
         {
             Item.width = 16;
             Item.height = 16;
-            Item.maxStack = 999;
+            Item.maxStack = 9999;
             Item.rare = ItemRarityID.LightRed;
 			Item.value = Item.buyPrice(0, 10);
         }